Add Detectron2 source
As this will be the main backbone of the project, including it seems to make the most sense, as it will need to be modified and extended for the rest of the LOFAR tasks.

Those tasks, as of now, seem to be for using either semantic segmentation or boxes around a optical and radio sources. Additionally, the source component identification is also a segmentation task.

There is also an idea for using extra auxiliary information such as the redshift and other known information. This would require a more thorough change to the network to accomodate that information though, so will be put off for later.