Speed up CNN generation, add more information
Now saves out the records as it goes incase the file needs to be regenerated, as currently it takes a long time for that to happen.

Adds the Angle that the source is at too, so that the sources with the same distance and same color can be distinguished as well, and should help with the source finding. Also saves the SkyCoord values so that if wanted, the Angles and such can be recomputed or rotated for data augmentation and the like, or converted to X-Y coordinates or whatever